+++ src/main/scala/com/dataintuitive/viash/platforms/NextflowLegacyPlatform.scala
@@ -27,7 +27,7 @@ import com.dataintuitive.viash.helpers.Circe._
 /**
  * / * Platform class for generating NextFlow (DSL2) modules.
  */
-case class NextFlowPlatform(
+case class NextflowLegacyPlatform(
   id: String = "nextflow",
   image: Option[String],
   tag: Option[Version] = None,
@@ -48,8 +48,9 @@ case class NextFlowPlatform(
   directive_time: Option[String] = None,
   directive_memory: Option[String] = None,
   directive_cache: Option[String] = None,
-  oType: String = "nextflow"
-) extends Platform {
+  oType: String = "nextflow",
+  variant: String = "legacy"
+) extends NextflowPlatform {
   assert(version.isEmpty, "nextflow platform: attribute 'version' is deprecated")
 
   private val nativePlatform = NativePlatform(id = id)